+++ tests/integration/helm_traces_enabled_test.go
@@ -130,12 +130,12 @@ func Test_Helm_Traces_Enabled(t *testing.T) {
 				waitDuration,
 				tickDuration,
 				stepfuncs.WithNameF(
-					stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrumentation"),
+					stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrument"),
 				),
 				stepfuncs.WithLabelsF(
 					stepfuncs.LabelFormatterKV{
 						K: "app",
-						V: stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrumentation"),
+						V: stepfuncs.ReleaseFormatter("%s-sumologic-otelcol-instrument"),
 					},
 				),
 			),